+++ src/main/java/com/honlife/core/app/controller/dashboard/payload/RoutineTotalCountResponse.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class RoutineTotalCountResponse {
     
     // 그 주 총 루틴 수
-    private Integer totalCount;
+    private Long totalCount;
     // 그 주 완료된 루틴 수
-    private Integer completedCount;
+    private Long completedCount;
 }
